+++ delivery/src/main/java/com/springcloud/client/delivery/application/DeliveryUpdateCommand.java
@@ -1,6 +1,7 @@
-package com.springcloud.client.delivery.domain.delivery;
+package com.springcloud.client.delivery.application;
 
 
+import com.springcloud.client.delivery.domain.delivery.DeliveryStatusEnum;
 import lombok.Builder;
 import lombok.Getter;
