+++ delivery/src/main/java/com/springcloud/client/delivery/application/IdentityIntegrationDTO.java
@@ -1,4 +1,4 @@
-package com.spring_cloud.eureka.client.order.domain.order;
+package com.springcloud.client.delivery.application;
 
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -8,7 +8,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.apache.kafka.common.serialization.Serializer;
 
-import java.util.List;
 import java.util.UUID;
 
 
@@ -16,13 +15,13 @@ import java.util.UUID;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class IdentityIntegrationResponse implements Serializer {
+public class IdentityIntegrationDTO implements Serializer {
 
     private static ObjectMapper objectMapper = new ObjectMapper();
     private UUID userId;
     private UUID hubId;
     private UUID companyId;
-    private List<UUID> orderId;
+    private UUID orderId;
     private UUID deliveryId;
 
     @Override
@@ -35,9 +34,5 @@ public class IdentityIntegrationResponse implements Serializer {
         } catch (Exception e) {
             throw new RuntimeException("Error serializing OrderCreateEvent", e);
         }
-    }
-
-    public List<UUID> getOrderIdList() {
-        return this.orderId;
     }
 }
