+++ delivery/src/main/java/com/springcloud/client/delivery/presentation/dto/DeliveryUpdateRequest.java
@@ -1,4 +1,7 @@
-package com.springcloud.client.delivery.domain.delivery;
+package com.springcloud.client.delivery.presentation.dto;
+
+import com.springcloud.client.delivery.application.DeliveryUpdateCommand;
+import com.springcloud.client.delivery.domain.delivery.DeliveryStatusEnum;
 
 import java.util.UUID;
