+++ order/src/main/java/com/spring_cloud/eureka/client/order/domain/order/OrderService.java
@@ -1,13 +1,16 @@
 package com.spring_cloud.eureka.client.order.domain.order;
 
 
+import com.spring_cloud.eureka.client.order.application.IdentityIntegrationCommand;
+import com.spring_cloud.eureka.client.order.application.OrderCreateCommand;
+import com.spring_cloud.eureka.client.order.application.OrderUpdateCommand;
 import com.spring_cloud.eureka.client.order.infrastructure.client.DeliveryClient;
 import com.spring_cloud.eureka.client.order.infrastructure.client.ProductClient;
 import com.spring_cloud.eureka.client.order.infrastructure.client.dto.*;
 import com.spring_cloud.eureka.client.order.infrastructure.client.dto.OrderCreateEvent;
 import com.spring_cloud.eureka.client.order.infrastructure.repository.OrderRepository;
 
-import com.spring_cloud.eureka.client.order.interfaces.OrderReadCommand;
+import com.spring_cloud.eureka.client.order.application.OrderReadCommand;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
