+++ order/src/main/java/com/spring_cloud/eureka/client/order/infrastructure/client/dto/IdentityIntegrationResponse.java
@@ -1,4 +1,4 @@
-package com.springcloud.client.delivery.domain.delivery;
+package com.spring_cloud.eureka.client.order.infrastructure.client.dto;
 
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -8,6 +8,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.apache.kafka.common.serialization.Serializer;
 
+import java.util.List;
 import java.util.UUID;
 
 
@@ -15,13 +16,13 @@ import java.util.UUID;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class IdentityIntegrationDTO implements Serializer {
+public class IdentityIntegrationResponse implements Serializer {
 
     private static ObjectMapper objectMapper = new ObjectMapper();
     private UUID userId;
     private UUID hubId;
     private UUID companyId;
-    private UUID orderId;
+    private List<UUID> orderId;
     private UUID deliveryId;
 
     @Override
@@ -34,5 +35,9 @@ public class IdentityIntegrationDTO implements Serializer {
         } catch (Exception e) {
             throw new RuntimeException("Error serializing OrderCreateEvent", e);
         }
+    }
+
+    public List<UUID> getOrderIdList() {
+        return this.orderId;
     }
 }
